+++ CHANGES.rst
@@ -5,6 +5,7 @@
 - Add search support to `~asdf.AsdfFile.schema_info`. [#1187]
 - Add `asdf.search.AsdfSearchResult` support for `~asdf.AsdfFile.schema_info` and
   `~asdf.search.AsdfSearchResult.schema_info` method. [#1197]
+- Use forc ndarray flag to correctly test for fortran array contiguity [#1206]
 
 2.13.0 (2022-08-19)
 -------------------